+++ src/components/Routes.js
@@ -5,6 +5,7 @@ import NavBar from '../containers/Navbar';
 import Home from '../containers/Home';
 import Cars from '../containers/Cars';
 import Reservation from '../containers/Reservation';
+import Login from '../containers/Login';
 import '../styles/App.css';
 import store from '../store';
 
@@ -16,7 +17,8 @@ const Routes = () => {
         <Switch>
           <Route exact path='/' component={Home} />
           {/* <Route exact path='/reservation' component={Reservation} />
-          <Route exact path='car' component={Cars}  /> */}
+          <Route exact path='/car' component={Cars}  /> */}
+          <Route exact path='/login' component={Login} />
         </Switch>
       </Router>
     </Provider>
